+++ src/main/java/com/WalkiePaw/security/oauth/CustomOAuth2User.java
@@ -16,7 +16,7 @@ public class CustomOAuth2User extends DefaultOAuth2User {
 
   private String email;
   private String nickname;
-  private Integer memberId;
+  private Long memberId;
   private String name;
   private String photoUrl;
   private Role role;
@@ -31,7 +31,7 @@ public class CustomOAuth2User extends DefaultOAuth2User {
    */
   public CustomOAuth2User(Collection<? extends GrantedAuthority> authorities,
       Map<String, Object> attributes, String nameAttributeKey,
-      String email, Integer memberId, String nickname, String name, String photoUrl, Role role) {
+      String email, Long memberId, String nickname, String name, String photoUrl, Role role) {
     super(authorities, attributes, nameAttributeKey);
     this.email = email;
     this.memberId = memberId;
